requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
email-validator
cryptography
